Create cards package and move cards.kt there
