+++ server/src/main/kotlin/core/pokerhands/Straight.kt
@@ -1,6 +1,6 @@
 package core.pokerhands
 
-import core.Card
+import core.cards.Card
 
 class Straight(cards: List<Card>) : PokerHand(HandRank.STRAIGHT, cards) {
